Research paper file
The main purpose of this research is to investigate the perception of  youth on the use of social media in politics. In this research, a comprehensive literature review has been conducted on research ideas based on similar content of the use of social media in politics. Five variables, such as voting behaviour, perceived usefulness, ease of use, political participation and personalization politics have been used to investigate the youth perception regarding the research objectives. In this research positivism research philosophy and descriptive approach has been followed for primary data collection and analysis. Quantitative statistical tools have been used to find out actual results from the study. Frequencies analysis, correlation,  mean, median and mode have been used to evaluate data. From the analysis, it has been found that the use of social media in politics has a robust impact on the political behaviour of youth on social media. It positively affective the voting behaviour of youth and pursue them to participate in various political activities. The finding of this study reveals that the use of social media in politics has a significant impact on the voting behaviour of people who are using social media in political communication. This research also found that politically interested people get needed information on social media and they get updated about political activities and new policies. The usefulness of social media has a positive impact on the use of social media by youth in political activities. This study also found that social media is easy to use and need less effort to perform communication activities. This pursues youth to use social media in political participation. Another finding is that the use of social media in politics increase youth participation in political activities. The traditional form of communication has a lack of access by the general people. But, digital media allow people to participate in decision making and criticism. This facilitates youth participation in political activities.
We would like to give our first deepest appreciation to Christine Tidåsen, who has supported us through this research by providing proper guidelines throughout our complete research process with unlimited patient and support. Then we would like to give our sincere thanks and respect to Professor Anders Pehrsson, who also supported by giving effective and informative guidelines and suggestions and his quality knowledge in the research process. Finally, we would like to provide thanks to our friends and family for their support and encouragement through this research process.
 Keywords: Social Media, Politics, Political Participation, Voting Behaviour, ease of use, perceived usefulness.